Scale Class

Base class for all scale types

Methods

Public methods

Scale$new()

Scale$transform()

Scale$render_axis()

Scale$clone()

Method new()

Initialize a new scale

Usage

Scale$new(aesthetic, name = NULL, breaks = NULL, labels = NULL, limits = NULL)

Arguments

aesthetic

The aesthetic this scale applies to (e.g., "x", "y", "color")

name

The name to use for the axis or legend

breaks

Vector of break points

labels

Vector of labels for the breaks

limits

Vector of length 2 giving the range of the scale

Method transform()

Transform data from data space to plot space

Usage

Scale$transform(x)

Arguments

x

Values to transform

Method render_axis()

Render the axis or legend for this scale

Usage

Scale$render_axis(side = "bottom")

Arguments

side

The side to render the axis on ("top", "right", "bottom", "left")

Method clone()

The objects of this class are cloneable with this method.

Usage

Scale$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.